+++ forui/lib/src/widgets/button/button.dart
@@ -122,6 +122,7 @@ class FButton extends StatelessWidget {
     final enabled = onPress != null || onLongPress != null;
 
     return FTappable.animated(
+      focusedOutlineStyle: style.focusedOutlineStyle,
       autofocus: autofocus,
       focusNode: focusNode,
       onFocusChange: onFocusChange,