+++ forui/lib/src/widgets/calendar/day/day_picker.dart
@@ -216,7 +216,11 @@ final class FCalendarDayPickerStyle with Diagnosticable {
         assert(0 < tileSize, 'tileSize must be positive.');
 
   /// Creates a [FCalendarDayPickerStyle] that inherits from the given [colorScheme] and [typography].
-  factory FCalendarDayPickerStyle.inherit({required FColorScheme colorScheme, required FTypography typography}) {
+  factory FCalendarDayPickerStyle.inherit({
+    required FColorScheme colorScheme,
+    required FTypography typography,
+    required FStyle style,
+  }) {
     final textStyle = typography.base.copyWith(color: colorScheme.foreground, fontWeight: FontWeight.w500);
     final mutedTextStyle = typography.base.copyWith(
       color: colorScheme.disable(colorScheme.mutedForeground),
@@ -227,14 +231,14 @@ final class FCalendarDayPickerStyle with Diagnosticable {
       selectedStyle: FCalendarEntryStyle(
         backgroundColor: colorScheme.primaryForeground,
         textStyle: mutedTextStyle,
-        focusedBorderColor: colorScheme.primaryForeground,
         radius: const Radius.circular(4),
+        focusedOutlineStyle: style.focusedOutlineStyle,
       ),
       unselectedStyle: FCalendarEntryStyle(
         backgroundColor: colorScheme.background,
         textStyle: mutedTextStyle,
-        focusedBorderColor: colorScheme.background,
         radius: const Radius.circular(4),
+        focusedOutlineStyle: style.focusedOutlineStyle,
       ),
     );
 
@@ -245,30 +249,30 @@ final class FCalendarDayPickerStyle with Diagnosticable {
           selectedStyle: FCalendarEntryStyle(
             backgroundColor: colorScheme.foreground,
             textStyle: typography.base.copyWith(color: colorScheme.background, fontWeight: FontWeight.w500),
-            focusedBorderColor: colorScheme.foreground,
             radius: const Radius.circular(4),
+            focusedOutlineStyle: style.focusedOutlineStyle,
           ),
           unselectedStyle: FCalendarEntryStyle(
             backgroundColor: colorScheme.background,
             textStyle: textStyle,
             hoveredBackgroundColor: colorScheme.secondary,
-            focusedBorderColor: colorScheme.foreground,
             radius: const Radius.circular(4),
+            focusedOutlineStyle: style.focusedOutlineStyle,
           ),
         ),
         enclosing: FCalendarDayStyle(
           selectedStyle: FCalendarEntryStyle(
             backgroundColor: colorScheme.primaryForeground,
             textStyle: mutedTextStyle,
-            focusedBorderColor: colorScheme.foreground,
             radius: const Radius.circular(4),
+            focusedOutlineStyle: style.focusedOutlineStyle,
           ),
           unselectedStyle: FCalendarEntryStyle(
             backgroundColor: colorScheme.background,
             textStyle: mutedTextStyle,
             hoveredBackgroundColor: colorScheme.secondary,
-            focusedBorderColor: colorScheme.foreground,
             radius: const Radius.circular(4),
+            focusedOutlineStyle: style.focusedOutlineStyle,
           ),
         ),
       ),